pyproject.toml
[project]
name = "quant_engine"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
  "numpy",
  "pandas",
  "scipy",
  "pyyaml",
]

[build-system]
requires = ["setuptools>=67", "wheel"]
build-backend = "setuptools.build_meta"

[project.scripts]
qe-run = "quant_engine.run_day:main"
qe-gt-align = "scripts.compare_alpha_gt:main" 